Separate out code to a greater extent
* Move relavent constants out to specific handler classes
* Create an encoder class
* Create a rotation (IR) handler class
* New main header for constants, to help sharing with other test files
* Reduce the amount of code in the main tasty-totems.cpp

+++ src/tasty-totems/lib/LedStrip/LedStrip.cpp
@@ -4,7 +4,7 @@
 
 #include "LedStrip.h"
 
-LedStrip::LedStrip(int16_t strip_pin, int num_pixels) {
+LedStrip::LedStrip(uint8_t strip_pin, int num_pixels) {
     _strip_pin = strip_pin;
     _num_pixels = num_pixels;
